+++ AFQ/nn/synthseg.py
@@ -74,7 +74,7 @@ def run_synthseg(t1_img, model_name):
 
     output_img = nbp.resample_from_to(
         nib.Nifti1Image(
-            output.astype(np.float32),
+            output.astype(np.uint8),
             t1_img_conformed.affine),
         t1_img)
